+++ src/device/local_data.cpp
@@ -6,14 +6,14 @@
 #include <logging.h>
 cpu procData[smp::max_cpu];
 
-extern "C" void asm_sse_save(uint64_t addr);
-extern "C" void asm_sse_load(uint64_t addr);
+extern "C" void asm_sse_save(uintptr_t addr);
+extern "C" void asm_sse_load(uintptr_t addr);
 void set_current_data(cpu *dat)
 {
 
     dat->me = dat;
     dat->current_processor_id = apic::the()->get_current_processor_id();
-    x86_wrmsr(LOCAL_DATA_DMSR, (uint64_t)dat);
+    x86_wrmsr(LOCAL_DATA_DMSR, (uintptr_t)dat);
 }
 void cpu::load_sse(uint64_t *data)
 {
